% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigrquerystorage.R
\docType{package}
\name{bigrquerystorage-package}
\alias{bigrquerystorage}
\alias{bigrquerystorage-package}
\title{bigrquerystorage: An Interface to Google's 'BigQuery Storage' API}
\description{
Easily talk to Google's 'BigQuery Storage' API from R (\url{https://docs.cloud.google.com/bigquery/docs/reference/storage/rpc}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/meztez/bigrquerystorage}
  \item Report bugs at \url{https://github.com/meztez/bigrquerystorage/issues}
}

}
\author{
\strong{Maintainer}: Bruno Tremblay \email{openr@neoxone.com}

Other contributors:
\itemize{
  \item  Google LLC [copyright holder, funder]
}

}
