% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{DBI}
\alias{DBI}
\alias{dbFetch,BigQueryResult-method}
\alias{dbReadTable,BigQueryConnection,character-method}
\title{DBI methods}
\usage{
\S4method{dbFetch}{BigQueryResult}(res, n = -1, ...)

\S4method{dbReadTable}{BigQueryConnection,character}(conn, name, ...)
}
\arguments{
\item{res}{An object inheriting from \linkS4class{DBIResult}, created by
\code{\link[DBI:dbSendQuery]{dbSendQuery()}}.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
or \code{n = Inf}
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{...}{Other arguments passed on to methods.}

\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[DBI:dbConnect]{dbConnect()}}.}

\item{name}{A character string specifying the unquoted DBMS table name,
or the result of a call to \code{\link[DBI:dbQuoteIdentifier]{dbQuoteIdentifier()}}.}
}
\description{
Implementations of pure virtual functions defined in the `DBI` package.
}
\keyword{internal}
