% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqs_download.R
\name{overload_bq_table_download}
\alias{overload_bq_table_download}
\title{Substitute bigrquery bq_table_download method. This is very experimental.}
\usage{
overload_bq_table_download(parent)
}
\arguments{
\item{parent}{Parent project used by the API for billing.}
}
\description{
Substitute bigrquery bq_table_download method. This is very experimental.
}
