% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigrquerystorage.R
\docType{package}
\name{bigrquerystorage-package}
\alias{bigrquerystorage-package}
\title{bigrquerystorage: R Client for BigQuery Storage API}
\description{
Provides facilities to interact with BigQuery Storage API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/meztez/bigrquerystorage}
  \item Report bugs at \url{https://github.com/meztez/bigrquerystorage/issues}
}

}
\author{
\strong{Maintainer}: Bruno Tremblay \email{openr@neoxone.com}

Other contributors:
\itemize{
  \item  Google LLC [copyright holder, funder]
}

}
