% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqs_download.R
\name{bqs_auth}
\alias{bqs_auth}
\alias{bqs_deauth}
\title{Initialize bigrquerystorage client}
\usage{
bqs_auth()

bqs_deauth()
}
\value{
No return value, called for side effects.
}
\description{
Initialize bigrquerystorage client

Close bigrquerystorage client
}
\details{
Will attempt to reuse \code{bigrquery} credentials.

About Credentials

If your application runs inside a Google Cloud environment that has
a default service account, your application can retrieve the service
account credentials to call Google Cloud APIs. Such environments
include Compute Engine, Google Kubernetes Engine, App Engine,
Cloud Run, and Cloud Functions. We recommend using this strategy
because it is more convenient and secure than manually passing credentials.

Additionally, we recommend you use Google Cloud Client Libraries for
your application. Google Cloud Client Libraries use a library called
Application Default Credentials (ADC) to automatically find your service
account credentials. ADC looks for service account credentials
in the following order:
\enumerate{
\item If the environment variable GOOGLE_APPLICATION_CREDENTIALS is set,
ADC uses the service account file that the variable points to.
\item If the environment variable GOOGLE_APPLICATION_CREDENTIALS isn't
set, ADC uses the default service account that Compute Engine,
Google Kubernetes Engine, App Engine, Cloud Run, and Cloud
Functions provide.
\item If ADC can't use either of the above credentials, an error occurs.
}
}
