% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqs_download.R
\name{bqs_table_download}
\alias{bqs_table_download}
\title{Download table from BigQuery using BigQuery Storage API}
\usage{
bqs_table_download(x, billing)
}
\arguments{
\item{x}{BigQuery table reference `{project}.{dataset}.{table_name}`}

\item{billing}{Billing project. Used as parent for `CreateReadSession` grpc method.}
}
\description{
Download table from BigQuery using BigQuery Storage API
}
