% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqs_download.R
\name{bqs_table_download}
\alias{bqs_table_download}
\title{Download table from BigQuery using BigQuery Storage API}
\usage{
bqs_table_download(
  x,
  parent = getOption("bigquerystorage.project", ""),
  snapshot_time = NA,
  selected_fields = character(),
  row_restriction = "",
  n_max = Inf,
  quiet = NA,
  as_tibble = FALSE,
  bigint = c("integer", "integer64", "numeric", "character"),
  max_results = lifecycle::deprecated()
)
}
\arguments{
\item{x}{BigQuery table reference `{project}.{dataset}.{table_name}`}

\item{parent}{Used as parent for `CreateReadSession`
grpc method. You can set option `bigquerystorage.project`.}

\item{snapshot_time}{Snapshot time}

\item{selected_fields}{A character vector of field to select from table.}

\item{row_restriction}{Restriction to apply to the table.}

\item{quiet}{Should information be printed to console.}

\item{as_tibble}{Should data be returned as tibble. Default is to return
as arrow Table from raw IPC stream.}

\item{bigint}{The R type that BigQuery's 64-bit integer types should be mapped to.
The default is `"integer"` which returns R's `integer` type but results in `NA` for
values above/below +/- 2147483647. `"integer64"` returns a [bit64::integer64],
which allows the full range of 64 bit integers.}

\item{max_results}{Maximum number of results to retrieve. Use `Inf` or `-1L`
retrieve all rows.}
}
\description{
Download table from BigQuery using BigQuery Storage API
}
